Fixed some bugs on import data Functions to edit the currency data Functions to edit the fuel data

+++ SolverClientComunication/Models/DbExchangeRates.cs
@@ -18,7 +18,7 @@ namespace SolverClientComunication.Models
         public string CurrencySymbol { get; set; }
 
         [Required]
-        public Decimal ValueInDolar { get; set; }
+        public Double ValueInDolar { get; set; }
 
         [Required]
         public DbInstance Instance { get; set; }
